overlay app-misc/ca-certificates: Add dependency on app-misc/c_rehash
It was relying on dev-libs/openssl pulling app-misc/c_rehash. But in
Gentoo, openssl dropped its dependency on c_rehash, as openssl does
not need it. Our ca-certificates package is using c_rehash in its
systemd service, so depend on c_rehash directly.

This will remove one modification we make to openssl package.